back/requirements.txt
asgiref==3.4.1backports.zoneinfo==0.2.1Django==4.0gunicorn==20.1.0django-filter==21.1djangorestframework==3.13.1psycopg2==2.9.2pytz==2021.3sqlparse==0.4.2tzdata==2021.5django-cors-headers==3.11.0drf-spectacular==0.22.1